Fixing import
I realised there was an error in this version I committed. I discovered that this version might not work as I performed an npm audit fix before I started working on the project (as several of the dependencies appeared as vulnerable) and this could have caused issues. The code is the same and you could see final product here: http://magento.ijewebservices.net/